service/requirements.txt
arrow==0.17.0
astropy==4.2
bcrypt==3.2.0
dask==2.30.0
distributed==2.30.1
fire==0.3.1
matplotlib==3.3.3
penquins>=2.0.0
pymongo==3.11.2
pyyaml==5.3.1
supervisor==4.2.1
